moved componets to plots folder
